I've added the importance check using the entropy method, I planned to do pruning also but I'm really out of time and efforts. I've changed the return type of getImportance to double(as it should be) also fixed a minor bug in the tree, it should be good now

+++ Attributes/Attribute.java
@@ -2,9 +2,8 @@ package Attributes;
 
 import java.util.List;
 
-import AttributeValues.AttributeValue;
-
 import util.Example;
+import AttributeValues.AttributeValue;
 
 /**
  * Interface for the attributes used to make decision trees
@@ -24,10 +23,12 @@ public interface Attribute {
 	/**
 	 * get the importance of this attribute in the example collection
 	 * 
+	 * Calculation of the importance is done using the binary entropy method
+	 * 
 	 * @param exampleCollection
-	 * @return
+	 * @return double value of the entropy deduction
 	 */
-	public int getImportance(List<Example> exampleCollection);
+	public double getImportance(List<Example> exampleCollection);
 
 	/**
 	 * get the value of the attribute in this example
